requirements.txt
flask==3.1.0
openai==1.68.0
pdfplumber==0.11.6
pypdf==5.4.0
reportlab==4.3.1
python-dotenv==1.1.0
werkzeug==3.1.3